Fix unit test file after renaming
